feat(job-store):データ登録時、すでにデータが存在した時にクライアントエラーを返すように - 見通しを良くするためにファイルをエンドポイントとアプリに分けた - jobInsetの中の処理が汚いが、一旦後で
